+++ engine/src/main/kotlin/world/gregs/voidps/engine/Script.kt
@@ -1,5 +1,10 @@
 package world.gregs.voidps.engine
 
+import com.github.michaelbull.logging.InlineLogger
+import kotlinx.coroutines.CoroutineExceptionHandler
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import world.gregs.voidps.engine.client.ui.dialogue.Dialogues
 import world.gregs.voidps.engine.client.variable.VariableApi
 import world.gregs.voidps.engine.data.SettingsReload
@@ -14,12 +19,26 @@ import world.gregs.voidps.engine.entity.character.player.skill.Skills
 import world.gregs.voidps.engine.inv.InventoryApi
 import world.gregs.voidps.engine.inv.Items
 import world.gregs.voidps.engine.timer.TimerApi
+import kotlin.coroutines.cancellation.CancellationException
 
 /**
  * A helper interface made up of all callable methods for easier scripting.
  */
 interface Script : Spawn, Despawn, Skills, Moved, VariableApi, TimerApi, Operation, Approachable, InterfaceInteraction, Death, SettingsReload, Dialogues, Items, InventoryApi, Hunt, Teleport, CombatApi {
     companion object {
+
+        private val logger = InlineLogger()
+        private val scope: CoroutineScope = CoroutineScope(Dispatchers.Unconfined)
+
+        fun launch(block: suspend CoroutineScope.() -> Unit) {
+            scope.launch(errorHandler, block = block)
+        }
+        private val errorHandler = CoroutineExceptionHandler { _, throwable ->
+            if (throwable !is CancellationException) {
+                logger.warn(throwable) { "Error in script handler." }
+            }
+        }
+
         val interfaces: MutableList<AutoCloseable> = mutableListOf(
             Spawn,
             Despawn,
