+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/instruction/InstructionHandlers.kt
@@ -11,7 +11,6 @@ import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.Players
 import world.gregs.voidps.engine.entity.item.floor.FloorItems
 import world.gregs.voidps.engine.entity.obj.GameObjects
-import world.gregs.voidps.engine.event.Event
 import world.gregs.voidps.engine.get
 import world.gregs.voidps.network.client.Instruction
 import world.gregs.voidps.network.client.instruction.*
@@ -76,7 +75,6 @@ class InstructionHandlers(
 
     fun handle(player: Player, instruction: Instruction) {
         when (instruction) {
-            is Event -> player.emit(instruction)
             is InteractInterfaceItem -> interactInterfaceItem.validate(player, instruction)
             is InteractInterfacePlayer -> interactInterfacePlayer.validate(player, instruction)
             is InteractInterfaceObject -> interactInterfaceObject.validate(player, instruction)
