+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/instruction/handle/ExecuteCommandHandler.kt
@@ -1,10 +1,10 @@
 package world.gregs.voidps.engine.client.instruction.handle
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.client.instruction.InstructionHandler
 import world.gregs.voidps.engine.client.command.Commands
 import world.gregs.voidps.engine.entity.character.move.tele
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.event.Events
 import world.gregs.voidps.engine.event.AuditLog
 import world.gregs.voidps.network.client.instruction.ExecuteCommand
 
@@ -28,7 +28,7 @@ class ExecuteCommandHandler : InstructionHandler<ExecuteCommand>() {
             player["world_map_marker_player"] = player.tile.id
             return
         }
-        Events.events.launch {
+        Script.launch {
             AuditLog.event(player, "command", "\"${instruction.command}\"")
             Commands.call(player, instruction.command)
         }
