+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/instruction/handle/InterfaceOptionHandler.kt
@@ -1,13 +1,13 @@
 package world.gregs.voidps.engine.client.instruction.handle
 
 import com.github.michaelbull.logging.InlineLogger
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.client.instruction.InstructionHandler
 import world.gregs.voidps.engine.client.instruction.InterfaceHandler
 import world.gregs.voidps.engine.client.ui.InterfaceOption
 import world.gregs.voidps.engine.data.definition.InterfaceDefinitions
 import world.gregs.voidps.engine.entity.InterfaceInteraction
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.event.Events
 import world.gregs.voidps.network.client.instruction.InteractInterface
 
 class InterfaceOptionHandler(
@@ -39,7 +39,7 @@ class InterfaceOptionHandler(
             item = item,
             itemSlot = itemSlot,
         )
-        Events.events.launch {
+        Script.launch {
             InterfaceInteraction.option(player, event)
         }
     }
