+++ engine/src/main/kotlin/world/gregs/voidps/engine/data/SettingsReload.kt
@@ -1,10 +1,6 @@
 package world.gregs.voidps.engine.data
 
 import it.unimi.dsi.fastutil.objects.ObjectArrayList
-import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.event.Event
-import world.gregs.voidps.engine.event.EventDispatcher
-import world.gregs.voidps.engine.event.Events
 
 interface SettingsReload {
