+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/InterfaceInteraction.kt
@@ -1,11 +1,11 @@
 package world.gregs.voidps.engine.entity
 
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.client.ui.InterfaceOption
 import world.gregs.voidps.engine.client.ui.ItemOption
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.item.Item
-import world.gregs.voidps.engine.event.Events
 import world.gregs.voidps.engine.event.Wildcard
 import world.gregs.voidps.engine.event.Wildcards
 
@@ -129,7 +129,7 @@ interface InterfaceInteraction {
         }
 
         fun itemOption(player: Player, option: String, item: Item = Item.EMPTY, slot: Int = 0, inventory: String = "inventory") {
-            Events.events.launch {
+            Script.launch {
                 itemOption(player, ItemOption(item, slot, inventory, option))
             }
         }
