+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/InterfaceFloorItemInteract.kt
@@ -1,12 +1,10 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.entity.item.floor.FloorItem
-import world.gregs.voidps.engine.entity.obj.GameObject
-import world.gregs.voidps.engine.event.Events
 
 data class InterfaceFloorItemInteract(
     override val target: FloorItem,
@@ -28,7 +26,7 @@ data class InterfaceFloorItemInteract(
     }
 
     private fun invoke(map: Map<String, List<suspend Player.(InterfaceFloorItemInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             for (block in map["$id:${target.id}"] ?: map["$id:*"] ?: map["*:${target.id}"] ?: return@launch) {
                 block(player, this@InterfaceFloorItemInteract)
             }
