+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/InterfaceNPCInteract.kt
@@ -1,10 +1,10 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.npc.NPC
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.event.Events
 
 data class InterfaceNPCInteract(
     override val target: NPC,
@@ -25,7 +25,7 @@ data class InterfaceNPCInteract(
     }
 
     private fun invoke(map: Map<String, List<suspend Player.(InterfaceNPCInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             for (block in map["$id:${target.def(player).stringId}"] ?: map["$id:*"] ?: map["*:${target.def(player).stringId}"] ?: return@launch) {
                 block(player, this@InterfaceNPCInteract)
             }
