+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/InterfaceObjectInteract.kt
@@ -1,11 +1,10 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.entity.obj.GameObject
-import world.gregs.voidps.engine.event.Events
 
 data class InterfaceObjectInteract(
     override val target: GameObject,
@@ -26,7 +25,7 @@ data class InterfaceObjectInteract(
     }
 
     private fun invoke(map: Map<String, List<suspend Player.(InterfaceObjectInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             for (block in map["$id:${target.def(player).stringId}"] ?: map["$id:*"] ?: map["*:${target.def(player).stringId}"] ?: return@launch) {
                 block(player, this@InterfaceObjectInteract)
             }
