+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/ItemFloorItemInteract.kt
@@ -1,12 +1,11 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.entity.item.floor.FloorItem
-import world.gregs.voidps.engine.entity.obj.GameObject
-import world.gregs.voidps.engine.event.Events
 
 data class ItemFloorItemInteract(
     override val target: FloorItem,
@@ -29,7 +28,7 @@ data class ItemFloorItemInteract(
     }
 
     private fun invoke(noDelays: Set<String>, map: Map<String, List<suspend Player.(ItemFloorItemInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             if (!noDelays.contains("${item.id}:${target.id}")) {
                 player.arriveDelay()
             }
