+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/ItemNPCInteract.kt
@@ -1,11 +1,11 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.npc.NPC
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.item.Item
-import world.gregs.voidps.engine.event.Events
 
 data class ItemNPCInteract(
     override val target: NPC,
@@ -27,7 +27,7 @@ data class ItemNPCInteract(
     }
 
     private fun invoke(map: Map<String, List<suspend Player.(ItemNPCInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             for (block in map["${item.id}:${target.def(player).stringId}"] ?: map["*:${target.def(player).stringId}"] ?: map["${item.id}:*"] ?: return@launch) { // Hack for spells
                 block(player, this@ItemNPCInteract)
             }
