+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/ItemObjectInteract.kt
@@ -1,11 +1,11 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.entity.obj.GameObject
-import world.gregs.voidps.engine.event.Events
 
 data class ItemObjectInteract(
     override val target: GameObject,
@@ -27,7 +27,7 @@ data class ItemObjectInteract(
     }
 
     private fun invoke(noDelays: Set<String>, map: Map<String, List<suspend Player.(ItemObjectInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             if (!noDelays.contains("${item.id}:${target.def(player).stringId}")) {
                 player.arriveDelay()
             }
