+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/NPCNPCInteract.kt
@@ -1,9 +1,9 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.npc.NPC
-import world.gregs.voidps.engine.event.Events
 
 data class NPCNPCInteract(
     override val target: NPC,
@@ -23,7 +23,7 @@ data class NPCNPCInteract(
     }
 
     private fun invoke(map: Map<String, List<suspend NPC.(NPCNPCInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             for (block in map[option] ?: return@launch) {
                 block(npc, this@NPCNPCInteract)
             }
