+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/NPCPlayerInteract.kt
@@ -1,10 +1,10 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.npc.NPC
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.event.Events
 
 data class NPCPlayerInteract(
     override val target: Player,
@@ -24,7 +24,7 @@ data class NPCPlayerInteract(
     }
 
     private fun invoke(map: Map<String, List<suspend NPC.(NPCPlayerInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             for (block in map[option] ?: return@launch) {
                 block(npc, this@NPCPlayerInteract)
             }
