+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/PlayerObjectInteract.kt
@@ -1,10 +1,10 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.obj.GameObject
-import world.gregs.voidps.engine.event.Events
 
 data class PlayerObjectInteract(
     override val target: GameObject,
@@ -25,7 +25,7 @@ data class PlayerObjectInteract(
     }
 
     private fun invoke(noDelays: Set<String>, map: Map<String, List<suspend Player.(PlayerObjectInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             val id = target.def(player).stringId
             if (!noDelays.contains("$option:$id") && (!noDelays.contains("$option:*") && !map.containsKey("$option:$id"))) {
                 player.arriveDelay()
