+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/PlayerPlayerInteract.kt
@@ -1,9 +1,9 @@
 package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.entity.Approachable
 import world.gregs.voidps.engine.entity.Operation
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.event.Events
 
 data class PlayerPlayerInteract(
     override val target: Player,
@@ -23,7 +23,7 @@ data class PlayerPlayerInteract(
     }
 
     private fun invoke(map: Map<String, List<suspend Player.(PlayerPlayerInteract) -> Unit>>) {
-        Events.events.launch {
+        Script.launch {
             for (block in map[option] ?: return@launch) {
                 block(player, this@PlayerPlayerInteract)
             }
