--- engine/src/main/kotlin/world/gregs/voidps/engine/event/TargetContext.kt
@@ -1,11 +1,0 @@
-package world.gregs.voidps.engine.event
-
-import world.gregs.voidps.engine.entity.Entity
-import world.gregs.voidps.engine.entity.character.Character
-
-/**
- * Context of an event, queue or action with a [target] entity
- */
-interface TargetContext<C : Character, T : Entity> : Context<C> {
-    val target: T
-}
