+++ engine/src/main/kotlin/world/gregs/voidps/engine/inv/transact/ChangeManager.kt
@@ -2,8 +2,6 @@ package world.gregs.voidps.engine.inv.transact
 
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.item.Item
-import world.gregs.voidps.engine.event.Event
-import world.gregs.voidps.engine.event.EventDispatcher
 import world.gregs.voidps.engine.inv.*
 import java.util.*
