+++ game/src/main/kotlin/content/entity/player/combat/special/SpecialAttack.kt
@@ -6,9 +6,6 @@ import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.entity.character.Character
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.equip.equipped
-import world.gregs.voidps.engine.event.Event
-import world.gregs.voidps.engine.event.EventDispatcher
-import world.gregs.voidps.engine.event.Events
 import world.gregs.voidps.network.login.protocol.visual.update.player.EquipSlot
 import kotlin.math.floor
