+++ game/src/test/kotlin/WorldTest.kt
@@ -38,7 +38,6 @@ import world.gregs.voidps.engine.entity.item.floor.FloorItems
 import world.gregs.voidps.engine.entity.obj.GameObject
 import world.gregs.voidps.engine.entity.obj.GameObjects
 import world.gregs.voidps.engine.entity.obj.ObjectShape
-import world.gregs.voidps.engine.event.Events
 import world.gregs.voidps.engine.event.Wildcards
 import world.gregs.voidps.engine.get
 import world.gregs.voidps.engine.inv.Inventory
@@ -253,7 +252,6 @@ abstract class WorldTest : KoinTest {
     fun afterAll() {
         saves?.deleteRecursively()
         ContentLoader.clear()
-        Events.events.clear()
         stopKoin()
     }
