+++ game/src/test/kotlin/content/activity/penguin_hide_and_seek/PenguinHideAndSeekTest.kt
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.assertNotNull
 import org.junit.jupiter.api.assertNull
 import world.gregs.voidps.engine.data.Settings
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
-import world.gregs.voidps.engine.event.Events
 import world.gregs.voidps.engine.inv.add
 import world.gregs.voidps.engine.inv.inventory
 import world.gregs.voidps.type.Tile
@@ -231,10 +230,7 @@ class PenguinHideAndSeekTest : WorldTest() {
     }
 
     private fun instance(): PenguinHideAndSeek {
-        val backup = Events.events
-        Events.setEvents(Events())
         val instance = PenguinHideAndSeek()
-        Events.setEvents(backup)
         return instance
     }
 }
